Rubocop Safe Autocorrect

+++ lib/tasks/delete_duplicate_orders.rake
@@ -1,31 +1,30 @@
-namespace :delete do
-  desc "delete duplicacy"
+# frozen_string_literal: true
 
-  task :duplicate_order, [:tenant] => :environment do |t, args|
+namespace :delete do
+  desc 'delete duplicacy'
+
+  task :duplicate_order, [:tenant] => :environment do |_t, args|
     tenants = []
-    if args[:tenant].present?
-      tenants = Tenant.where(name: args[:tenant])
-    else
-      tenants = Tenant.all
-    end
+    tenants = if args[:tenant].present?
+                Tenant.where(name: args[:tenant])
+              else
+                Tenant.all
+              end
     tenants.each do |tenant|
-      begin
-        Apartment::Tenant.switch!(tenant.name)
-        Order.all.group_by(&:increment_id).each do |key, orders|
-          next if orders.count == 1
-          scanned_true = ((orders.map(&:status).include? ("scanned")) || (orders.map(&:status).include? ("cancelled")))
-          if scanned_true
-            orders.each do |dup_order|
-              dup_order.destroy if !(dup_order.status == "scanned" || dup_order.status == "cancelled")
-            end
-          else
-            orders.drop(1).each do |dup_order|
-              dup_order.destroy
-            end
+      Apartment::Tenant.switch!(tenant.name)
+      Order.all.group_by(&:increment_id).each do |_key, orders|
+        next if orders.count == 1
+
+        scanned_true = ((orders.map(&:status).include? 'scanned') || (orders.map(&:status).include? 'cancelled'))
+        if scanned_true
+          orders.each do |dup_order|
+            dup_order.destroy unless dup_order.status == 'scanned' || dup_order.status == 'cancelled'
           end
+        else
+          orders.drop(1).each(&:destroy)
         end
-      rescue
       end
+    rescue StandardError
     end
   end
 end
